Update heartrate postprocessor to use vectorization.
This commit introduces two new postprocessors, CustomMaskPostprocessor and HeartRateInterpolationPostprocessor, to improve the handling of masked regions and heart rate data interpolation in the LSTM experiment. The BaseMhcDataset class is updated to utilize these postprocessors, allowing for more robust data preparation. Additionally, the LSTM experiment script is modified to reflect these changes, including adjustments to the use of masked loss during training. Unit tests are updated to ensure the correct application of the new processing logic and maintain data integrity.